services/transcription-agent/requirements.txt
websockets
nats-py
asyncio